requirements.txt
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fonttools==4.61.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
ma==1.0.2
ma.mongo==1.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
msgpack==1.1.2
numba==0.63.0
numpy==2.3.5
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.45
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.1
Werkzeug==3.1.4
